PLAT-533: use message instead of execution
