+++ ledger-core/virtual/integration/method_test.go
@@ -16,7 +16,6 @@ import (
 
 	"github.com/insolar/assured-ledger/ledger-core/application/builtin/proxy/testwallet"
 	testwalletProxy "github.com/insolar/assured-ledger/ledger-core/application/builtin/proxy/testwallet"
-	"github.com/insolar/assured-ledger/ledger-core/application/testwalletapi/statemachine"
 	"github.com/insolar/assured-ledger/ledger-core/insolar"
 	"github.com/insolar/assured-ledger/ledger-core/insolar/contract"
 	"github.com/insolar/assured-ledger/ledger-core/insolar/payload"
@@ -29,7 +28,6 @@ import (
 	"github.com/insolar/assured-ledger/ledger-core/testutils/gen"
 	"github.com/insolar/assured-ledger/ledger-core/testutils/runner/logicless"
 	"github.com/insolar/assured-ledger/ledger-core/vanilla/throw"
-	"github.com/insolar/assured-ledger/ledger-core/virtual/authentication"
 	"github.com/insolar/assured-ledger/ledger-core/virtual/descriptor"
 	"github.com/insolar/assured-ledger/ledger-core/virtual/execute"
 	"github.com/insolar/assured-ledger/ledger-core/virtual/integration/utils"
@@ -1747,197 +1745,6 @@ func TestVirtual_CallContractTwoTimes(t *testing.T) {
 
 	require.Equal(t, 4, typedChecker.VCallRequest.Count())
 	require.Equal(t, 2, typedChecker.VCallResult.Count())
-
-	mc.Finish()
-}
-
-// b.Foo calls a.Bar, pulse changes, outgoing has proper token
-func TestVirtual_CallMethodOutgoing_WithPulseChange(t *testing.T) {
-	t.Log("C5141")
-
-	mc := minimock.NewController(t)
-
-	server, ctx := utils.NewUninitializedServer(nil, t)
-	defer server.Stop()
-
-	logger := inslogger.FromContext(ctx)
-
-	executeDone := server.Journal.WaitStopOf(&execute.SMExecute{}, 2)
-
-	runnerMock := logicless.NewServiceMock(ctx, mc, func(execution execution.Context) string {
-		return execution.Request.CallSiteMethod
-	})
-	server.ReplaceRunner(runnerMock)
-
-	authService := authentication.NewServiceMock(mc)
-
-	authService.GetCallDelegationTokenMock.Set(func(outgoing reference.Global, to reference.Global, pn pulse.Number, object reference.Global) (c1 payload.CallDelegationToken) {
-		service := authentication.NewService(ctx, server.JetCoordinatorMock)
-		return service.GetCallDelegationToken(
-			outgoing, to, pn, object)
-	})
-	authService.HasToSendTokenMock.Return(true) // always send tokens
-	// skip token checks, in tests we have only one node, so we need to allow self signed tokens
-	authService.IsMessageFromVirtualLegitimateMock.Set(func(_ context.Context, payloadObj interface{}, sender reference.Global, pr pulse.Range) (mustReject bool, err error) {
-		return false, nil
-	})
-
-	server.ReplaceAuthenticationService(authService)
-
-	server.Init(ctx)
-	server.IncrementPulseAndWaitIdle(ctx)
-
-	typedChecker := server.PublisherMock.SetTypedChecker(ctx, mc, server)
-
-	var (
-		barIsolation = contract.MethodIsolation{
-			Interference: contract.CallTolerable,
-			State:        contract.CallDirty,
-		}
-		fooIsolation = contract.MethodIsolation{
-			Interference: contract.CallTolerable,
-			State:        contract.CallDirty,
-		}
-
-		classA        = gen.UniqueGlobalRef()
-		objectAGlobal = reference.NewSelf(server.RandomLocalWithPulse())
-
-		objectBGlobal = reference.NewSelf(server.RandomLocalWithPulse())
-
-		callOutgoing = server.RandomLocalWithPulse()
-
-		outgoingCallRef = reference.NewRecordOf(objectBGlobal, callOutgoing)
-	)
-
-	Method_PrepareObject(ctx, server, payload.Ready, objectAGlobal)
-	Method_PrepareObject(ctx, server, payload.Ready, objectBGlobal)
-
-	// add ExecutionMocks to runnerMock
-	{
-		outgoingCall := execution.NewRPCBuilder(outgoingCallRef, objectBGlobal).
-			CallMethod(objectAGlobal, classA, "Bar", []byte("123")).
-			SetInterference(barIsolation.Interference).
-			SetIsolation(barIsolation.State)
-		objectBExecutionMock := runnerMock.AddExecutionMock("Foo")
-		objectBExecutionMock.AddStart(
-			func(ctx execution.Context) {
-				logger.Debug("ExecutionStart [B.Foo]")
-				assert.Equal(t, objectBGlobal, ctx.Object)
-				assert.Equal(t, callOutgoing, ctx.Request.CallOutgoing)
-			},
-			&execution.Update{
-				Type:     execution.OutgoingCall,
-				Error:    nil,
-				Outgoing: outgoingCall,
-			},
-		)
-		objectBExecutionMock.AddContinue(
-			func(result []byte) {
-				logger.Debug("ExecutionContinue [B.Foo]")
-				require.Equal(t, []byte("finish A.Bar"), result)
-			},
-			&execution.Update{
-				Type:   execution.Done,
-				Result: requestresult.New([]byte("finish B.Foo"), objectBGlobal),
-			},
-		)
-
-		runnerMock.AddExecutionMock("Bar").AddStart(
-			func(ctx execution.Context) {
-				logger.Debug("ExecutionStart [A.Bar]")
-				assert.Equal(t, objectAGlobal, ctx.Request.Callee)
-				assert.Equal(t, objectBGlobal, ctx.Request.Caller)
-				assert.Equal(t, []byte("123"), ctx.Request.Arguments)
-			},
-			&execution.Update{
-				Type:   execution.Done,
-				Result: requestresult.New([]byte("finish A.Bar"), objectBGlobal),
-			},
-		)
-
-		runnerMock.AddExecutionClassify("Bar", barIsolation, nil)
-		runnerMock.AddExecutionClassify("Foo", fooIsolation, nil)
-	}
-
-	callCount := 0
-
-	firstPulse := server.GetPulse().PulseNumber
-
-	{
-		typedChecker.VStateReport.SetResend(true)
-		typedChecker.VDelegatedCallRequest.SetResend(true)
-		typedChecker.VDelegatedCallResponse.SetResend(true)
-		typedChecker.VDelegatedRequestFinished.SetResend(true)
-		typedChecker.VCallRequest.Set(func(request *payload.VCallRequest) bool {
-			assert.Equal(t, objectAGlobal, request.Callee)
-			assert.Equal(t, objectBGlobal, request.Caller)
-			assert.Equal(t, []byte("123"), request.Arguments)
-			assert.Equal(t, payload.CTMethod, request.CallType)
-			assert.Equal(t, uint32(1), request.CallSequence)
-			assert.Equal(t, outgoingCallRef, request.CallReason)
-			assert.Equal(t, payload.BuildCallFlags(barIsolation.Interference, barIsolation.State), request.CallFlags)
-
-			assert.Equal(t, firstPulse, request.CallOutgoing.Pulse())
-
-			if callCount == 0 {
-				assert.Equal(t, payload.CallDelegationToken{}, request.DelegationSpec)
-				server.IncrementPulseAndWaitIdle(ctx)
-				callCount++
-				// request will be sent in previous pulse
-				// ommit sending
-				return false
-			}
-
-			service := authentication.NewService(ctx, server.JetCoordinatorMock)
-			origToken := service.GetCallDelegationToken(
-				outgoingCallRef, server.GlobalCaller(), server.GetPulse().PulseNumber, objectBGlobal)
-
-			assert.Equal(t, origToken, request.DelegationSpec)
-
-			return true // resend
-		})
-		typedChecker.VCallResult.Set(func(res *payload.VCallResult) bool {
-			assert.Equal(t, payload.BuildCallFlags(fooIsolation.Interference, fooIsolation.State), res.CallFlags)
-
-			switch res.Callee {
-			case objectAGlobal:
-				assert.Equal(t, []byte("finish A.Bar"), res.ReturnArguments)
-				assert.Equal(t, payload.CTMethod, res.CallType)
-				assert.Equal(t, objectBGlobal, res.Caller)
-			case objectBGlobal:
-				assert.Equal(t, []byte("finish B.Foo"), res.ReturnArguments)
-				assert.Equal(t, payload.CTMethod, res.CallType)
-				assert.Equal(t, statemachine.APICaller, res.Caller)
-				assert.Equal(t, int(firstPulse), int(res.CallOutgoing.Pulse()))
-			default:
-				t.Fatal("wrong Callee")
-			}
-			// we should resend that message only if it's CallResult from outgoing
-			return res.Caller == objectBGlobal
-		})
-	}
-
-	pl := payload.VCallRequest{
-		CallType:       payload.CTMethod,
-		CallFlags:      payload.BuildCallFlags(fooIsolation.Interference, fooIsolation.State),
-		Caller:         statemachine.APICaller,
-		Callee:         objectBGlobal,
-		CallSiteMethod: "Foo",
-		CallOutgoing:   callOutgoing,
-	}
-
-	msg := server.WrapPayload(&pl).SetSender(statemachine.APICaller).Finalize()
-	server.SendMessage(ctx, msg)
-
-	// wait for all calls and SMs
-	testutils.WaitSignalsTimed(t, 10*time.Second, executeDone)
-	testutils.WaitSignalsTimed(t, 10*time.Second, server.Journal.WaitAllAsyncCallsDone())
-
-	assert.Equal(t, 2, typedChecker.VCallRequest.Count())
-	assert.Equal(t, 2, typedChecker.VCallResult.Count())
-	assert.Equal(t, 2, typedChecker.VStateReport.Count())
-	assert.Equal(t, 1, typedChecker.VDelegatedCallRequest.Count())
-	assert.Equal(t, 1, typedChecker.VDelegatedCallResponse.Count())
 
 	mc.Finish()
 }
